backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
anthropic
python-dotenv
numpy
pydantic